+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - changed: Adjust Transaction List Scene spacings, remove "Transactions" header
 - changed: Add a referral ID to share links
 - changed: Use `EdgeAsset` for defaultWallets in app config
+- fixed: Fix button placement on wallet activation scene
 - fixed: Resolved levitating search bar bug on Android
 - fixed: Show a useful, localized error message when device doesn't have an email account
 - fixed: Make FilledTextInputs take up constant vertical space