+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - added: Notification status analytics param
 - added: @bugsnag/react-native-performance
 - fixed: Some text inputs not selectable for some Samsung devices
+- fixed: "Mini" button incorrect max width layout
 
 ## 4.1.0